refactor(pin-input): updated component structure
