+++ packages/react/src/components/pin-input/index.ts
@@ -1,0 +1,11 @@
+export {
+  PinInputField,
+  PinInputPropsContext,
+  PinInputRoot,
+  usePinInputPropsContext,
+} from "./pin-input"
+export type { PinInputFieldProps, PinInputRootProps } from "./pin-input"
+
+export * from "./pin-input.style"
+
+export * as PinInput from "./namespace"
