+++ packages/react/src/components/pin-input/namespace.ts
@@ -1,0 +1,10 @@
+export {
+  PinInputField as Field,
+  PinInputPropsContext as PropsContext,
+  PinInputRoot as Root,
+  usePinInputPropsContext as usePropsContext,
+} from "./pin-input"
+export type {
+  PinInputFieldProps as FieldProps,
+  PinInputRootProps as RootProps,
+} from "./pin-input"
