+++ packages/react/src/index.ts
@@ -77,7 +77,7 @@ export * from "./components/number-input"
 export * from "./components/password-input"
 export * from "./components/picture"
 export * from "./components/picture"
-// export * from "./components/pin-input"
+export * from "./components/pin-input"
 // export * from "./components/popover"
 export * from "./components/portal"
 export * from "./components/reorder"
