--- packages/react/src/legacy-components/pin-input/index.ts
@@ -1,2 +1,0 @@
-export { PinInput, PinInputField } from "./pin-input"
-export type { PinInputFieldProps, PinInputProps } from "./pin-input"
